+++ src/modules/user/user.service.ts
@@ -69,8 +69,8 @@ export class UserService {
         const refreshToken = await this.jwtManagerService.generateRefreshToken({ login });
 
         await this.redisService.setTokens(login, accessToken, refreshToken);
-        res.cookie('accessToken', accessToken, { httpOnly: true, sameSite: 'none', secure: false });
-        res.cookie('refreshToken', refreshToken, { httpOnly: true, sameSite: 'none', secure: false });
+        res.cookie('accessToken', accessToken, { httpOnly: true, sameSite: 'none', secure: true });
+        res.cookie('refreshToken', refreshToken, { httpOnly: true, sameSite: 'none', secure: true });
 
         const message = `User "${login}" has been successfully logged in.`;
         this.loggerService.info(context, message);
@@ -103,14 +103,14 @@ export class UserService {
         }
 
         const newAccessToken = await this.jwtManagerService.generateAccessToken(userPayload);
-        res.cookie('accessToken', newAccessToken, { httpOnly: true, sameSite: 'none', secure: false });
+        res.cookie('accessToken', newAccessToken, { httpOnly: true, sameSite: 'none', secure: true });
         this.loggerService.info(context, `Access token was renewed for ${userPayload.login} user.`);
         const refreshTokenPayload = await this.jwtManagerService.verifyRefreshToken(refreshToken);
         let newRefreshToken = null;
 
         if (isTooShortToExpireRefreshToken(refreshTokenPayload)) {
             newRefreshToken = await this.jwtManagerService.generateRefreshToken(refreshTokenPayload);
-            res.cookie('refreshToken', newRefreshToken, { httpOnly: true, sameSite: 'none', secure: false });
+            res.cookie('refreshToken', newRefreshToken, { httpOnly: true, sameSite: 'none', secure: true });
             this.loggerService.info(context, `Refresh token was renewed for ${userPayload.login} user.`);
         }