+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.0] - 2024-11-18
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): New ingredient category elements.
+- [Marek Kurańda](https://github.com/mjkuranda): `mushrooms` and `cereal-products` as a new ingredient categories.
+- [Marek Kurańda](https://github.com/mjkuranda): List for `pantry ingredients`.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): Ingredient elements within each category.
+- [Marek Kurańda](https://github.com/mjkuranda): Path to dish images from `/data/images/meals` to `/data/images/dishes`
+- [Marek Kurańda](https://github.com/mjkuranda): Update for `axios` and `nodemailer` packages.
+
 ## [2.5.1] - 2024-10-21
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): `API_URL/users/:activationCode/activate` to `API_URL/users/activate/:activationCode` within email message.